+++ src/lib.rs
@@ -1,2 +1,6 @@
 #![no_std]
 #![allow(dead_code)]
+
+mod resistance;
+
+pub use resistance::{ExtF32 as ResistanceExtF32, ExtU32 as ResistanceExtU32, Resistance};