+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.0] - 2023-02-02
+## [0.1.0] - 2023-02-02
 
 ### Added
 
-- Placeholder for library code
+- `Resistance` struct for storing resistance values
+- `Resistance` extension methods for `u32` and `f32` values